requirements.txt
Flask==2.3.2
yt-dlp==2024.*
google-cloud-speech==2.*
google-cloud-storage==2.*
youtube-transcript-api==0.4.0
google-cloud-firestore==2.*